smart_engine/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
tritonclient[grpc]==2.41.0
numpy==1.26.3
opencv-python==4.9.0.80
requests==2.31.0
python-dotenv==1.0.0
SQLAlchemy==2.0.25
alembic==1.13.1
pydantic==2.5.3
fastapi>=0.95.0
uvicorn>=0.21.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
pydantic-settings==2.1.0
mysqlclient==2.2.1 
redis>=4.5.1
tritonclient[grpc]>=2.30.0
numpy>=1.23.0
opencv-python>=4.5.0 